Created nav-bar and sorted login/out functionality
