+++ ses/src/main/webapp/WEB-INF/views/logout-success.jsp
@@ -3,17 +3,18 @@
 <!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
 <html>
 <head>
-	<link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.0/css/bootstrap.min.css">
 	<meta http-equiv="Content-Type" content="text/html; charset=ISO-8859-1">
-	<title>stockSim - Admin Home</title>
+	<link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.0/css/bootstrap.min.css">
+	<title>stockSim - Logout</title>
 </head>
-<body>
-	<jsp:include page="../nav.jsp"></jsp:include>
+<body>	
+	<jsp:include page="nav.jsp"/>
+	
+	<br/>
 	<div class="container">
-		<br/>
-		<h1>ADMIN HOME</h1>
-		<br/>
-		<a href="/"><input type="button" value="Home"/></a>
+		<h1>You have been logged out</h1>
+		<p>See you again soon!</p>
+		<a href="/login"><button type="button" class="btn btn-primary">Back to Login</button></a>
 	</div>
 </body>
 </html>
