+++ ses/src/main/webapp/WEB-INF/views/nav-not-logged-in.jsp
@@ -1,0 +1,10 @@
+<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"%>
+
+<ul class="navbar-nav mr-auto">
+	<li class="nav-item">
+		<a class="nav-link" href="/login">Login</a>
+	</li>
+	<li class="nav-item">
+		<a class="nav-link" href="/register">Register</a>
+	</li>
+</ul>
