+++ ses/src/main/webapp/WEB-INF/views/user/hi.jsp
@@ -7,6 +7,7 @@
 <title>Demo Index Page</title>
 </head>
 <body>
+	<jsp:include page="../nav.jsp"/>
 	Hi hi :D
 	<br/>
 	User count: ${userCount}
