+++ ses/src/main/webapp/WEB-INF/views/user/myAccount.jsp
@@ -1,0 +1,28 @@
+<%@ page language="java" contentType="text/html; charset=ISO-8859-1"
+    pageEncoding="ISO-8859-1"%>
+<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
+<html>
+<head>
+	<meta http-equiv="Content-Type" content="text/html; charset=ISO-8859-1">
+	<link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.0/css/bootstrap.min.css">
+	<title>Demo Index Page</title>
+</head>
+<body>	
+	<jsp:include page="../nav.jsp"/>
+	
+	<br/>
+	<div class="container">
+		<h1>My Account</h1>
+		<h2>User Details</h2>
+		
+		
+		<br/>
+		<br/>
+		<br/>
+		<a href="hello"><button type="button" class="btn btn-primary">Some links</button></a>
+		<br/><br/>
+		<a href="/user/stockExchange"><button type="button" class="btn btn-primary">Stock Exchange</button></a>
+	</div>
+	
+</body>
+</html>
